Fix compiler warnings

+++ src/secure_utils.c
@@ -171,9 +171,9 @@ void unlock_page(void* base, size_t region_size) {
 
 void fisher_yates_shuffle(uint16_t *indicies, size_t length) {
 
-    for (int i = 0; i < length; i++) indicies[i] = i;
+    for (size_t i = 0; i < length; i++) indicies[i] = i;
 
-    for (int i = length - 1; i > 0; i--) {
+    for (size_t i = length - 1; i > 0; i--) {
         int j = _generate_random_number(i);
 
         int tmp = indicies[i];
